11/8/25 - 9:37am : Update
Did a bit of cleaning to the file systems. Started patching together what I had in my file with the repo. There are still a few dependencies broken but that is simply because Godot through Steam makes it a real pain to work with the files.
Anyways, this is also a test for the first push into the repo. Hope all goes well. :)